Debian 13 support

+++ distributions/base-compile-rocksdb.sh
@@ -1,5 +1,5 @@
-PIP3="pip3" 
-APT="apt"
+PIP3="${PIP3:-pip3}"
+APT="${APT:-apt}"
 
 function install_rocksdb {
 	if ! declare -f install_rocksdb_dependencies > /dev/null; then
@@ -9,6 +9,10 @@ function install_rocksdb {
 	_DIR=$(pwd)
 	
 	_info "Installing RocksDB... This will take a while."
-	$APT install librocksdb-dev
-	$PIP3 install python-rocksdb
+	$APT install -y librocksdb-dev
+	if declare -f pip_cmd > /dev/null; then
+		pip_cmd install python-rocksdb
+	else
+		$PIP3 install python-rocksdb
+	fi
 }
